refactor(mcp): extract hotel JSON-schema builders into schema_hotels.go
tools_hotels.go was 807 LOC — over the 800-LOC file-hygiene gate and a
responsibility collage. Move the five pure hotel schema/input-property builders
(hotelSearchOutputSchema, hotelPricesOutputSchema, hotelReviewsOutputSchema,
hotelRoomsOutputSchema, hotelSearchInputProperties) verbatim into a new
same-package file mcp/schema_hotels.go, along the schema/orchestration seam.
Pure move: no renames, no signature or behaviour change; callers unchanged.
tools_hotels.go: 807 -> 590 LOC. Both mcp/tools_flights.go (#458) and now
tools_hotels.go are under the gate.

Implemented by grok 4.5; reviewed by Claude. go build + go test ./mcp green,
no duplicate defs.
